vectorDB/scripts/requirements.txt
os
uuid
tiktoken
langchain
langchain-experimental
langchain-openai
pinecone
tqdm
vectorDB
python-dotenv
datetime
